+++ src/Symfony/Component/Form/Extension/Core/Type/TextType.php
@@ -12,10 +12,24 @@
 namespace Symfony\Component\Form\Extension\Core\Type;
 
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\DataTransformerInterface;
+use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class TextType extends AbstractType
+class TextType extends AbstractType implements DataTransformerInterface
 {
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        // When empty_data is explicitly set to an empty string,
+        // a string should always be returned when NULL is submitted
+        // This gives more control and thus helps preventing some issues
+        // with PHP 7 which allows type hinting strings in functions
+        // See https://github.com/symfony/symfony/issues/5906#issuecomment-203189375
+        if ('' === $options['empty_data']) {
+            $builder->addViewTransformer($this);
+        }
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -32,5 +46,23 @@ class TextType extends AbstractType
     public function getBlockPrefix()
     {
         return 'text';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function transform($data)
+    {
+        // Model data should not be transformed
+        return $data;
+    }
+
+    /**
+     * {@inheritdoc}
+     *.
+     */
+    public function reverseTransform($data)
+    {
+        return null === $data ? '' : $data;
     }
 }